refactor(scheduler): rename manager.py to service.py, keep shim for compat (#341)
Move SchedulerManager from src/scheduler/manager.py to src/scheduler/service.py.
Old path becomes a thin backward-compat shim. Update all src/, test, and doc imports.
