+++ tests/routes/test_pipelines_routes.py
@@ -10,7 +10,7 @@ from httpx import ASGITransport, AsyncClient
 from src.config import AppConfig
 from src.database import Database
 from src.models import Account, Channel
-from src.scheduler.manager import SchedulerManager
+from src.scheduler.service import SchedulerManager
 from src.search.ai_search import AISearchEngine
 from src.search.engine import SearchEngine
 from src.telegram.auth import TelegramAuth
