+++ tests/routes/test_scheduler_routes.py
@@ -12,7 +12,7 @@ from src.collection_queue import CollectionQueue
 from src.config import AppConfig
 from src.database import Database
 from src.models import Account, CollectionTaskStatus, SearchQuery, StatsAllTaskPayload
-from src.scheduler.manager import SchedulerManager
+from src.scheduler.service import SchedulerManager
 from src.search.ai_search import AISearchEngine
 from src.search.engine import SearchEngine
 from src.telegram.auth import TelegramAuth
