+++ tests/test_coverage_final.py
@@ -1038,7 +1038,7 @@ class TestDeepagentsSyncRemainingTools:
 
 class TestSchedulerSyncJobState:
     async def test_sync_job_disable(self):
-        from src.scheduler.manager import SchedulerManager
+        from src.scheduler.service import SchedulerManager
         mgr = SchedulerManager()
         mgr._scheduler = MagicMock()
         mgr._scheduler.running = True
@@ -1047,7 +1047,7 @@ class TestSchedulerSyncJobState:
         mgr._scheduler.remove_job.assert_called_once_with("collect")
 
     async def test_sync_job_enable_collection(self):
-        from src.scheduler.manager import SchedulerManager
+        from src.scheduler.service import SchedulerManager
         mgr = SchedulerManager()
         mgr._scheduler = MagicMock()
         mgr._scheduler.running = True
@@ -1057,7 +1057,7 @@ class TestSchedulerSyncJobState:
         mgr._scheduler.add_job.assert_called_once()
 
     async def test_sync_job_enable_photo_due(self):
-        from src.scheduler.manager import SchedulerManager
+        from src.scheduler.service import SchedulerManager
         mgr = SchedulerManager()
         mgr._scheduler = MagicMock()
         mgr._scheduler.running = True
@@ -1066,7 +1066,7 @@ class TestSchedulerSyncJobState:
         mgr._scheduler.add_job.assert_called_once()
 
     async def test_sync_job_enable_photo_auto(self):
-        from src.scheduler.manager import SchedulerManager
+        from src.scheduler.service import SchedulerManager
         mgr = SchedulerManager()
         mgr._scheduler = MagicMock()
         mgr._scheduler.running = True
@@ -1075,7 +1075,7 @@ class TestSchedulerSyncJobState:
         mgr._scheduler.add_job.assert_called_once()
 
     async def test_sync_job_enable_sq_prefix(self):
-        from src.scheduler.manager import SchedulerManager
+        from src.scheduler.service import SchedulerManager
         mgr = SchedulerManager()
         mgr._scheduler = MagicMock()
         mgr._scheduler.running = True
@@ -1084,7 +1084,7 @@ class TestSchedulerSyncJobState:
             mgr.sync_search_query_jobs.assert_called_once()
 
     async def test_sync_job_enable_pipeline_prefix(self):
-        from src.scheduler.manager import SchedulerManager
+        from src.scheduler.service import SchedulerManager
         mgr = SchedulerManager()
         mgr._scheduler = MagicMock()
         mgr._scheduler.running = True
@@ -1093,7 +1093,7 @@ class TestSchedulerSyncJobState:
             mgr.sync_pipeline_jobs.assert_called_once()
 
     async def test_get_job_next_run_fallback(self):
-        from src.scheduler.manager import SchedulerManager
+        from src.scheduler.service import SchedulerManager
         mgr = SchedulerManager()
         job = SimpleNamespace(id="test_job", next_run_time="2025-01-01")
         mgr._scheduler = MagicMock()
@@ -1103,7 +1103,7 @@ class TestSchedulerSyncJobState:
         assert result == "2025-01-01"
 
     async def test_get_all_jobs_cache(self):
-        from src.scheduler.manager import SchedulerManager
+        from src.scheduler.service import SchedulerManager
         mgr = SchedulerManager()
         mgr._scheduler = MagicMock()
         mgr._scheduler.get_jobs = MagicMock(return_value=[])
